refactor import statements and add new models for Line, User, Location, DeviceType, KnowledgeBase, Comment, and MacAddressHistory
