+++ app/auth/routes.py
@@ -13,7 +13,8 @@
 from flask import Blueprint, render_template, request, redirect, url_for, flash
 from flask_login import login_user, logout_user, login_required, current_user
 from werkzeug.security import generate_password_hash, check_password_hash
-from models import db, User
+from app.extensions import db
+from app.models import User
 
 auth_bp = Blueprint('auth', __name__)
