+++ app/models/__init__.py
@@ -1,5 +1,18 @@
 from app.extensions import db
-from app.models.auth import models
-from app.models.inventory import models
-from app.models.knowledge_base import models
-from app.models.work import models
+from .DeviceName import DeviceName
+from .DeviceType import DeviceType
+from .ForceDataHistory import ForceDataHistory
+from .KnowledgeBase import KnowledgeBase
+from .Line import Line
+from .Location import Location
+from .MacAddressHistory import MacAddressHistory
+from .ModuleHistory import ModuleHistory
+from .SerialNumberHistory import SerialNumberHistory
+from .User import User
+from .Work import Work
+
+
+
+
+
+
